+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-cosmos/src/main/java/com/microsoft/azure/toolkit/intellij/cosmos/code/function/CosmosDBDatabaseNameCompletionProvider.java
@@ -8,7 +8,6 @@ package com.microsoft.azure.toolkit.intellij.cosmos.code.function;
 import com.intellij.codeInsight.completion.CompletionParameters;
 import com.intellij.codeInsight.completion.CompletionProvider;
 import com.intellij.codeInsight.completion.CompletionResultSet;
-import com.intellij.codeInsight.completion.CompletionUtilCore;
 import com.intellij.codeInsight.lookup.LookupElement;
 import com.intellij.codeInsight.lookup.LookupElementBuilder;
 import com.intellij.openapi.module.Module;
@@ -19,6 +18,7 @@ import com.intellij.patterns.PsiJavaPatterns;
 import com.intellij.psi.PsiAnnotation;
 import com.intellij.psi.PsiAnnotationParameterList;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiLiteralExpression;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.util.ProcessingContext;
 import com.microsoft.azure.toolkit.ide.common.icon.AzureIcons;
@@ -30,14 +30,20 @@ import com.microsoft.azure.toolkit.intellij.connector.code.function.FunctionAnno
 import com.microsoft.azure.toolkit.intellij.connector.code.function.FunctionAnnotationTypeHandler;
 import com.microsoft.azure.toolkit.intellij.connector.code.function.FunctionAnnotationValueInsertHandler;
 import com.microsoft.azure.toolkit.intellij.connector.code.function.FunctionUtils;
+import com.microsoft.azure.toolkit.intellij.connector.dotazure.AzureModule;
 import com.microsoft.azure.toolkit.intellij.cosmos.connection.SqlCosmosDBAccountResourceDefinition;
 import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.auth.AzureAccount;
+import com.microsoft.azure.toolkit.lib.common.operation.OperationBundle;
+import com.microsoft.azure.toolkit.lib.common.telemetry.AzureTelemeter;
+import com.microsoft.azure.toolkit.lib.common.telemetry.AzureTelemetry;
+import com.microsoft.azure.toolkit.lib.cosmos.sql.SqlCosmosDBAccount;
 import com.microsoft.azure.toolkit.lib.cosmos.sql.SqlDatabase;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.*;
 
 import static com.intellij.patterns.PlatformPatterns.psiElement;
@@ -53,7 +59,7 @@ public class CosmosDBDatabaseNameCompletionProvider extends CompletionProvider<C
                             return StringUtils.equalsAnyIgnoreCase(psiAnnotation.getQualifiedName(), COSMOS_ANNOTATIONS);
                         }
                     })));
-    public static final PsiElementPattern<?, ?> COSMOS_DATABASE_PATTERN = psiElement().withSuperParent(2, COSMOS_DATABASE_NAME_PAIR_PATTERN);
+    public static final PsiElementPattern<?, ?> COSMOS_DATABASE_PATTERN = psiElement().withParent(PsiLiteralExpression.class).withSuperParent(2, COSMOS_DATABASE_NAME_PAIR_PATTERN);
 
     static {
         FunctionAnnotationTypeHandler.registerKeyPairPattern(COSMOS_DATABASE_NAME_PAIR_PATTERN);
@@ -69,18 +75,19 @@ public class CosmosDBDatabaseNameCompletionProvider extends CompletionProvider<C
             return;
         }
         final String connectionValue = FunctionUtils.getConnectionValueFromAnnotation(annotation);
-        final Object resource = Optional.ofNullable(FunctionUtils.getConnectionFromAnnotation(annotation))
-                .map(Connection::getResource).map(Resource::getData).orElse(null);
-        final SqlDatabase account = resource instanceof SqlDatabase ? (SqlDatabase) resource : null;
-        if (Objects.isNull(account) && StringUtils.isNotBlank(connectionValue)) {
+        final Connection<?, ?> connection = FunctionUtils.getConnectionFromAnnotation(annotation);
+        final SqlDatabase database = (SqlDatabase) Optional.ofNullable(connection).map(Connection::getResource)
+                .map(Resource::getData).filter(data -> data instanceof SqlDatabase).orElse(null);
+        if (Objects.isNull(database) && StringUtils.isNotBlank(connectionValue)) {
             return;
         }
-        final String fullPrefix = StringUtils.substringBefore(element.getText(), CompletionUtilCore.DUMMY_IDENTIFIER_TRIMMED).replace("\"", "").trim();
-        final List<SqlDatabase> accountsToSearch = Objects.nonNull(account) ? List.of(account) :
-                Utils.getConnectedResources(module, SqlCosmosDBAccountResourceDefinition.INSTANCE);
+        final List<SqlDatabase> accountsToSearch = Objects.nonNull(database) ? List.of(database) :
+            AzureModule.from(module).getConnections(SqlCosmosDBAccountResourceDefinition.INSTANCE).stream()
+                .filter(Connection::isValidConnection).map(Connection::getResource).map(Resource::getData).toList();
         accountsToSearch.stream()
-                .map(database -> createLookupElement(database, module))
+                .map(d -> createLookupElement(d, module))
                 .forEach(result::addElement);
+        AzureTelemeter.log(AzureTelemetry.Type.OP_END, OperationBundle.description("boundary/connector.complete_cosmos_database"));
     }
 
     private LookupElement createLookupElement(@Nonnull final SqlDatabase database, Module module) {
@@ -94,5 +101,20 @@ public class CosmosDBDatabaseNameCompletionProvider extends CompletionProvider<C
                 .withTypeText(database.getResourceTypeName())
                 .withTailText(" " + database.getParent().getName())
                 .withIcon(IntelliJAzureIcons.getIcon(AzureIcons.Cosmos.MODULE));
+    }
+
+    @Nullable
+    public static SqlDatabase getConnectedDatabase(@Nullable final PsiAnnotation annotation) {
+        if (Objects.isNull(annotation) || Objects.isNull(annotation.findAttribute("databaseName"))) {
+            return null;
+        }
+        final Connection<?, ?> connection = FunctionUtils.getConnectionFromAnnotation(annotation);
+        final SqlDatabase database = (SqlDatabase) Optional.ofNullable(connection).map(Connection::getResource)
+                .map(Resource::getData).filter(data -> data instanceof SqlDatabase).orElse(null);
+        if (Objects.isNull(database)) {
+            return null;
+        }
+        final String databaseName = Optional.ofNullable(annotation.findAttributeValue("databaseName")).map(PsiElement::getText).map(text -> text.replace("\"", "")).orElse(StringUtils.EMPTY);
+        return ((SqlCosmosDBAccount) database.getParent()).sqlDatabases().get(databaseName, database.getResourceGroupName());
     }
 }