+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-database-java/src/main/java/com/microsoft/azure/toolkit/intellij/database/connection/SpringSqlDatabaseResourceDefinition.java
@@ -10,13 +10,14 @@ import com.microsoft.azure.toolkit.intellij.connector.spring.SpringSupported;
 import com.microsoft.azure.toolkit.lib.database.entity.IDatabase;
 import org.apache.commons.lang3.tuple.Pair;
 
+import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
 
 public interface SpringSqlDatabaseResourceDefinition<T extends IDatabase> extends SpringSupported<T> {
 
     @Override
-    default public List<Pair<String, String>> getSpringProperties() {
+    default public List<Pair<String, String>> getSpringProperties(@Nullable final String key) {
         final List<Pair<String, String>> properties = new ArrayList<>();
         properties.add(Pair.of("spring.datasource.url", String.format("${%s_URL}", Connection.ENV_PREFIX)));
         properties.add(Pair.of("spring.datasource.username", String.format("${%s_USERNAME}", Connection.ENV_PREFIX)));