+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/code/AnnotationFixes.java
@@ -1,12 +1,16 @@
 package com.microsoft.azure.toolkit.intellij.connector.code;
 
 import com.intellij.codeInsight.intention.IntentionAction;
+import com.intellij.codeInspection.ProblemHighlightType;
 import com.intellij.codeInspection.util.IntentionFamilyName;
 import com.intellij.codeInspection.util.IntentionName;
+import com.intellij.lang.annotation.AnnotationHolder;
+import com.intellij.lang.annotation.HighlightSeverity;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleUtil;
 import com.intellij.openapi.project.Project;
+import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import com.intellij.util.IncorrectOperationException;
 import com.microsoft.azure.toolkit.intellij.connector.Connection;
@@ -17,9 +21,12 @@ import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.auth.AzureAccount;
 import com.microsoft.azure.toolkit.lib.common.action.Action;
 import com.microsoft.azure.toolkit.lib.common.action.AzureActionManager;
+import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
+import org.apache.commons.lang3.StringUtils;
 
 import javax.annotation.Nonnull;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.Consumer;
 
 public class AnnotationFixes {
@@ -29,6 +36,10 @@ public class AnnotationFixes {
     };
 
     public static IntentionAction createNewConnection(ResourceDefinition<?> definition, Consumer<Connection<?, ?>> callback) {
+        return createNewConnection(definition, callback, null);
+    }
+
+    public static IntentionAction createNewConnection(ResourceDefinition<?> definition, Consumer<Connection<?, ?>> callback, String defaultEnvPrefix) {
         return new IntentionAction() {
 
             @Override
@@ -49,12 +60,14 @@ public class AnnotationFixes {
             }
 
             @Override
+            @AzureOperation("user/connector.create_connection_quick_fix")
             public void invoke(@Nonnull Project project, Editor editor, PsiFile file) throws IncorrectOperationException {
                 final Module module = ModuleUtil.findModuleForFile(file);
                 if (Objects.nonNull(module)) {
                     final var dialog = new ConnectorDialog(project);
                     dialog.setConsumer(new ModuleResource(module.getName()));
                     dialog.setResourceDefinition(definition);
+                    Optional.ofNullable(defaultEnvPrefix).filter(StringUtils::isNoneBlank).ifPresent(dialog::setEnvPrefix);
                     if (dialog.showAndGet()) {
                         callback.accept(dialog.getValue());
                     } else {
@@ -92,6 +105,7 @@ public class AnnotationFixes {
             }
 
             @Override
+            @AzureOperation("user/connector.sign_in_quick_fix")
             public void invoke(@Nonnull Project project, Editor editor, PsiFile file) throws IncorrectOperationException {
                 if (!Azure.az(AzureAccount.class).isLoggedIn()) {
                     AzureActionManager.getInstance().getAction(Action.REQUIRE_AUTH).handle((a) -> callback.run());
@@ -105,4 +119,12 @@ public class AnnotationFixes {
         };
     }
 
+
+    public static void createSignInAnnotation(@Nonnull PsiElement element, @Nonnull AnnotationHolder holder) {
+        holder.newAnnotation(HighlightSeverity.WEAK_WARNING, "You are not signed in to Azure")
+                .range(element.getTextRange())
+                .highlightType(ProblemHighlightType.WEAK_WARNING)
+                .withFix(AnnotationFixes.signIn(AnnotationFixes.DO_NOTHING))
+                .create();
+    }
 }