+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/code/EditConnectionFix.java
@@ -14,22 +14,22 @@ import com.intellij.util.IncorrectOperationException;
 import com.microsoft.azure.toolkit.intellij.connector.Connection;
 import com.microsoft.azure.toolkit.intellij.connector.ResourceConnectionActionsContributor;
 import com.microsoft.azure.toolkit.lib.common.action.AzureActionManager;
-import com.microsoft.azure.toolkit.lib.common.model.AzResource;
+import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import lombok.AllArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 
 @AllArgsConstructor
 public class EditConnectionFix implements IntentionAction {
-    private final Connection<? extends AzResource, ?> connection;
+    private final Connection<?, ?> connection;
 
     @Override
     public @IntentionName @NotNull String getText() {
-        return "Edit Connection";
+        return "Edit connection";
     }
 
     @Override
     public @NotNull @IntentionFamilyName String getFamilyName() {
-        return "Resource Connection Fixes";
+        return "Resource connection fixes";
     }
 
     @Override
@@ -38,6 +38,7 @@ public class EditConnectionFix implements IntentionAction {
     }
 
     @Override
+    @AzureOperation("user/connector.edit_connection_quick_fix")
     public void invoke(@NotNull Project project, Editor editor, PsiFile file) throws IncorrectOperationException {
         final DataContext context = dataId -> CommonDataKeys.PROJECT.getName().equals(dataId) ? project : null;
         final AnActionEvent event = AnActionEvent.createFromAnAction(new EmptyAction(), null, "azure.annotator.action", context);