+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/code/spring/PropertiesValueCompletionProvider.java
@@ -32,7 +32,6 @@ import com.microsoft.azure.toolkit.lib.common.messager.ExceptionNotification;
 import com.microsoft.azure.toolkit.lib.common.model.AzResource;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.common.operation.OperationBundle;
-import com.microsoft.azure.toolkit.lib.common.task.AzureTaskManager;
 import com.microsoft.azure.toolkit.lib.common.telemetry.AzureTelemeter;
 import com.microsoft.azure.toolkit.lib.common.telemetry.AzureTelemetry;
 import lombok.RequiredArgsConstructor;
@@ -40,6 +39,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
@@ -78,6 +78,7 @@ public class PropertiesValueCompletionProvider extends CompletionProvider<Comple
         return definitions.stream().flatMap(d -> Utils.listResourceForDefinition(module.getProject(), d).stream().map(r -> LookupElementBuilder.create(r, r.getName())
             .withIcon(IntelliJAzureIcons.getIcon(StringUtils.firstNonBlank(r.getDefinition().getIcon(), AzureIcons.Common.AZURE.getIconPath())))
             .bold()
+            .withCaseSensitivity(false)
             .withLookupStrings(Arrays.asList(r.getName(), ((AzResource) r.getData()).getResourceGroupName()))
             .withInsertHandler(new PropertyValueInsertHandler(r))
             .withTailText(" " + ((AzResource) r.getData()).getResourceTypeName())
@@ -88,7 +89,7 @@ public class PropertiesValueCompletionProvider extends CompletionProvider<Comple
         final List<ResourceDefinition<?>> definitions = ResourceManager.getDefinitions(ResourceDefinition.RESOURCE).stream()
             .filter(d -> d instanceof SpringSupported<?>).toList();
         return definitions.stream().map(d -> (SpringSupported<?>) d)
-            .filter(d -> d.getSpringProperties().stream().anyMatch(p -> p.getKey().equals(key)))
+            .filter(d -> d.getSpringProperties(key).stream().anyMatch(p -> p.getKey().equals(key)))
             .toList();
     }
 
@@ -102,34 +103,29 @@ public class PropertiesValueCompletionProvider extends CompletionProvider<Comple
         @ExceptionNotification
         @AzureOperation(name = "user/connector.select_resource_completion_item_in_properties")
         public void handleInsert(@Nonnull InsertionContext context, @Nonnull LookupElement lookupElement) {
-            final PsiElement element = context.getFile().findElementAt(context.getStartOffset());
-            if (Objects.nonNull(element)) {
-                context.getDocument().deleteString(element.getTextOffset(), element.getTextOffset() + element.getTextLength());
-            }
+            context.getDocument().deleteString(context.getStartOffset(), context.getTailOffset());
             final Project project = context.getProject();
             final Module module = ModuleUtil.findModuleForFile(context.getFile().getVirtualFile(), project);
-            AzureTaskManager.getInstance().write(() -> Optional.ofNullable(module).map(AzureModule::from)
+            Optional.ofNullable(module).map(AzureModule::from)
                 .map(AzureModule::initializeWithDefaultProfileIfNot).map(Profile::getConnectionManager)
                 .ifPresent(connectionManager -> connectionManager
                     .getConnectionsByConsumerId(module.getName()).stream()
                     .filter(c -> Objects.equals(resource, c.getResource())).findAny()
                     .ifPresentOrElse(c -> insert(c, context),
-                        () -> Utils.createAndInsert(module, resource, context, connectionManager,
-                            PropertyValueInsertHandler::insert, PropertyValueInsertHandler::cancel))));
-        }
-
-        @AzureOperation(name = "user/connector.cancel_creating_connection_in_properties")
-        private static void cancel(@Nonnull InsertionContext context) {
+                        () -> connectionManager.getProfile().getModule().connect(resource, c -> insert(c, context))));
         }
 
         @AzureOperation(name = "user/connector.insert_value_in_properties")
-        public static void insert(Connection<?, ?> c, @Nonnull InsertionContext context) {
+        public static void insert(@Nullable Connection<?, ?> c, @Nonnull InsertionContext context) {
             final PsiElement element = context.getFile().findElementAt(context.getStartOffset());
-            final List<Pair<String, String>> properties = SpringSupported.getProperties(c);
-            if (properties.size() < 1 || Objects.isNull(element)) {
+            if (Objects.isNull(element) || Objects.isNull(c)) {
                 return;
             }
             final String k0 = element.getParent().getFirstChild().getText().trim();
+            final List<Pair<String, String>> properties = SpringSupported.getProperties(c, k0);
+            if (properties.size() < 1) {
+                return;
+            }
             properties.stream().filter(p -> p.getKey().equals(k0)).findAny().ifPresent(p -> {
                 properties.remove(p);
                 properties.add(0, p);