+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/code/spring/YamlKeyCompletionProvider.java
@@ -24,6 +24,9 @@ import com.microsoft.azure.toolkit.intellij.common.IntelliJAzureIcons;
 import com.microsoft.azure.toolkit.intellij.connector.dotazure.ResourceManager;
 import com.microsoft.azure.toolkit.intellij.connector.spring.SpringSupported;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
+import com.microsoft.azure.toolkit.lib.common.operation.OperationBundle;
+import com.microsoft.azure.toolkit.lib.common.telemetry.AzureTelemeter;
+import com.microsoft.azure.toolkit.lib.common.telemetry.AzureTelemetry;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Triple;
 import org.jetbrains.yaml.YAMLUtil;
@@ -54,11 +57,12 @@ public class YamlKeyCompletionProvider extends CompletionProvider<CompletionPara
         ResourceManager.getDefinitions().stream()
                 .filter(d -> d instanceof SpringSupported<?>)
                 .map(d -> (SpringSupported<?>) d)
-                .flatMap(d -> d.getSpringProperties().stream().map(p -> Triple.of(p.getKey(), p.getValue(), d)))
+                .flatMap(d -> d.getSpringProperties(key).stream().map(p -> Triple.of(p.getKey(), p.getValue(), d)))
                 .filter(t -> !StringUtils.startsWith(t.getLeft(), "#")) // filter out commented properties
                 .filter(t -> StringUtils.isBlank(key) || (StringUtils.startsWith(t.getLeft(), key) && !StringUtils.equals(t.getLeft(), key)))
                 .filter(t -> YAMLUtil.getQualifiedKeyInFile(yamlFile, getKeyListForProperty(t.getLeft())) == null)
                 .forEach(t -> result.addElement(createYamlKeyLookupElement(t.getLeft(), t.getRight(), yamlFile)));
+        AzureTelemeter.log(AzureTelemetry.Type.OP_END, OperationBundle.description("boundary/connector.complete_keys_in_yaml"));
     }
 
     private LookupElement createYamlKeyLookupElement(@Nonnull final String property, @Nonnull final SpringSupported<?> definition,
@@ -87,8 +91,10 @@ public class YamlKeyCompletionProvider extends CompletionProvider<CompletionPara
         context.commitDocument();
         YamlUtils.insertYamlKeyValue(property, null, yamlFile, context);
         final YAMLKeyValue result = YAMLUtil.getQualifiedKeyInFile(yamlFile, getKeyListForProperty(property));
-        context.getEditor().getCaretModel().moveToOffset(result.getTextRange().getEndOffset() + 1);
-        AutoPopupController.getInstance(context.getProject()).scheduleAutoPopup(context.getEditor());
+        if (Objects.nonNull(result)) {
+            context.getEditor().getCaretModel().moveToOffset(result.getTextRange().getEndOffset() + 1);
+            AutoPopupController.getInstance(context.getProject()).scheduleAutoPopup(context.getEditor());
+        }
     }
 
     private String[] getKeyListForProperty(@Nonnull final String property) {